fixed deletelink not working, launchlink actually launches things

+++ Tilandis/Tilandis.cpp
@@ -7,6 +7,7 @@
 // Standard headers
 #include <iostream>
 #include <string>
+#include <fstream>
 #include <tchar.h>
 #include <Windows.h>
 
@@ -22,10 +23,12 @@ int main() {
 		std::cerr << "Failed to prepare the link document." << std::endl;
 		return 1;
 	}
-	if (Tilandis::UsingCommandLine(__argc, __argv)) {
+	if (Tilandis::UsingCommandLine(__argc, __argv)) { // FIXME: Commandline processing, while it works, is extremely friggin weird because I wrote it stoned
 		try {
+			if (Tilandis::DeleteMode && Tilandis::CreateMode) { throw Tilandis::Exceptions::BadArgCombo; }
 			if (Tilandis::DeleteMode) { Tilandis::Links::DeleteLink(); }
-			else { Tilandis::Links::CreateLink(); }
+			if (Tilandis::CreateMode) { Tilandis::Links::CreateLink(); }
+			if (Tilandis::AddToRegistry) { Tilandis::RegisterProtocol(); }
 		}
 		catch (Tilandis::Exceptions::BadCommandLine exc) {
 			std::cerr << exc.what() << std::endl;
@@ -36,7 +39,18 @@ int main() {
 		if (__argc == 1) {
 			Tilandis::PrintUsage();
 		}
-		else if (__argc == 2) { Tilandis::Links::LaunchLink(__argv[1]); }
+		else if (__argc == 2) {
+			std::string LinkName = __argv[1];
+			size_t pos;
+			if ((pos = LinkName.find(":")) != std::string::npos) { // if we FIND a colon
+				LinkName.erase(0, pos + 1); // Remove protocols from the beginning
+			}
+			if (!Tilandis::Links::LaunchLink(LinkName.c_str())) {
+				std::ofstream err;
+				err.open("tilandis.err");
+				err << LinkName;
+			}
+		}
 	}
 	
 	return 0;
@@ -54,4 +68,12 @@ void Tilandis::PrintUsage() {
 	std::cout << "\t-a <args> :: Arguments to pass to the program." << std::endl;
 	std::cout << "\t-w <directory> :: The working directory to launch the program in. Defaults to the same as the file location." << std::endl;
 	std::cout << "\t-f :: Replace any existing link with the same name. (Mnemonic: \"force\")" << std::endl;
+}
+
+bool Tilandis::RegisterProtocol() {
+	HKEY registry;
+	DWORD regresult;
+	long WINAPI result = RegCreateKeyEx(HKEY_CLASSES_ROOT, Tilandis::RegistryProtocolName.c_str(), 0, NULL, REG_OPTION_NON_VOLATILE, KEY_ALL_ACCESS, NULL, &registry, &regresult);
+	if (regresult == REG_OPENED_EXISTING_KEY) { std::cout << "note: this protocol's already registered with something"; }
+	return true;
 }
